Added ability to add URL parameters to confirmation
There may be situations where you want to pass URL parameters along with the confirmation. In my use case, I wanted to add a parameter that would trigger animated confetti to be displayed along with the confirmation.

To add parameters, add CSS classes in this format urlparam-parameterkey-parameter-value.

For example, if I wanted to add "success=1" as a URL parameter, I would add this CSS class to the form settings: urlparam-success-1.

You should be able to add however many URL parameters that might be needed, but I've yet to actually test that.